Rewrote in older version, plus added more stuff
Redid points and outlines and started on the pcb. Added column_net fields to all the keys, but need to do that differently due to having 20 keys and the microcontroller having 18 pins that can be used for keys. Also need to add through-holes, reset-button-footprint, battery-connector-footprint, power-button-footprint. Might need to remove two keys, but would rather not.